+++ EasyAccomod.Core/Services/RequestExtends/RequestExtendService.cs
@@ -23,8 +23,17 @@ namespace EasyAccomod.Core.Services.RequestExtends
             this.context = context;
             this.userManager = userManager;
         }
-        public async Task<RequestExtend> RequestExtendPost(RequestExtendModel model)
+        public async Task<bool> CheckUserAndRole(long accessId, string role)
         {
+            var user = userManager.Users.Where(x => x.Id == accessId && x.IsConfirm).FirstOrDefault();
+            if (user == null) throw new ServiceException("Tài khoản không tồn tại");
+
+            return await userManager.IsInRoleAsync(user, role);
+        }
+            public async Task<RequestExtend> RequestExtendPost(RequestExtendModel model)
+        {
+            if (await CheckUserAndRole(model.UserId, CommonConstants.MODERATOR) == false && await CheckUserAndRole(model.UserId, CommonConstants.ADMIN) == false && await CheckUserAndRole(model.UserId, CommonConstants.OWNER) == false)
+                throw new ServiceException("Tài khoản không có quyền truy cập");
             var post = context.Posts.Where(x => x.PostId == model.PostId && x.PostStatus == PostStatusEnum.Accepted).FirstOrDefault();
             if (post == null) throw new ServiceException("Bài đăng không tồn tại hoặc chưa được confirm");
             var user = userManager.Users.Where(u => u.Id == model.UserId && u.IsConfirm).FirstOrDefault();
@@ -37,16 +46,30 @@ namespace EasyAccomod.Core.Services.RequestExtends
                 RequsetExtendStatus = RequsetExtendStatusEnum.Request,
                 UserName = user.UserName
             };
+            Notification notification = new Notification()
+            {
+                Content = CommonConstants.REQUEST_EXTEND,
+                NotifTime = DateTime.Now,
+                OfMod = true,
+                UserName = user.UserName,
+                PostId = post.PostId,
+                IsDelete = false
+            };
+            await context.Notifications.AddAsync(notification);
             await context.RequestExtends.AddAsync(requestExtend);
             await context.SaveChangesAsync();
             return requestExtend;
         }
-        public List<RequestExtend> GetPostsRequestExtend()
+        public async Task<List<RequestExtend>> GetPostsRequestExtend(long accessId)
         {
+            if (await CheckUserAndRole(accessId, CommonConstants.MODERATOR) == false && await CheckUserAndRole(accessId, CommonConstants.ADMIN) == false && await CheckUserAndRole(accessId, CommonConstants.OWNER) == false)
+                throw new ServiceException("Tài khoản không có quyền truy cập");
             return context.RequestExtends.Where(x => x.RequsetExtendStatus == RequsetExtendStatusEnum.Request).ToList();
         }
-        public async Task<Post> ConfirmRequestExtend(long requestId)
+        public async Task<Post> ConfirmRequestExtend(long requestId,long accessId)
         {
+            if (await CheckUserAndRole(accessId, CommonConstants.MODERATOR) == false && await CheckUserAndRole(accessId, CommonConstants.ADMIN) == false)
+                throw new ServiceException("Tài khoản không có quyền truy cập");
             var request = context.RequestExtends.Where(x => x.Id == requestId && x.RequsetExtendStatus == RequsetExtendStatusEnum.Request).FirstOrDefault();
             if (request == null) throw new ServiceException(@"Request không tồn tại hoặc đã được accept/reject");
             var post = context.Posts.Where(x => x.PostId == request.PostId && x.IsDetele == false).FirstOrDefault();
@@ -67,8 +90,10 @@ namespace EasyAccomod.Core.Services.RequestExtends
             await context.SaveChangesAsync();
             return post;
         }
-        public async Task<Post> RejectRequestExtend(long requestId)
+        public async Task<Post> RejectRequestExtend(long requestId,long accessId)
         {
+            if (await CheckUserAndRole(accessId, CommonConstants.MODERATOR) == false && await CheckUserAndRole(accessId, CommonConstants.ADMIN) == false )
+                throw new ServiceException("Tài khoản không có quyền truy cập");
             var request = context.RequestExtends.Where(x => x.Id == requestId && x.RequsetExtendStatus == RequsetExtendStatusEnum.Request).FirstOrDefault();
             if (request == null) throw new ServiceException(@"Request không tồn tại hoặc đã được accept/reject");
             var post = context.Posts.Where(x => x.PostId == request.PostId && x.IsDetele == false).FirstOrDefault();
@@ -89,8 +114,10 @@ namespace EasyAccomod.Core.Services.RequestExtends
             return post;
         }
 
-        public async Task<RequestExtend> DeleteRequestExtend(long requestId)
+        public async Task<RequestExtend> DeleteRequestExtend(long requestId,long accessId)
         {
+            if (await CheckUserAndRole(accessId, CommonConstants.MODERATOR) == false && await CheckUserAndRole(accessId, CommonConstants.ADMIN) == false)
+                throw new ServiceException("Tài khoản không có quyền truy cập");
             var request = context.RequestExtends.Where(x => x.Id == requestId).FirstOrDefault();
             if (request == null) throw new ServiceException(@"Request không tồn tại hoặc đã được accept/reject");
             context.RequestExtends.Remove(request);
